# docs/swagger.yaml
definitions:
...
  group.Group:
    properties:
      description:
        type: string
      name:
        type: string
      permissions:
        items:
          type: integer
        type: array
      sub_groups:
        items:
          type: integer
        type: array
    required:
    - name
    type: object
  group.GroupWindow:
    properties:
      groups_id:
        items:
          type: integer
        type: array
      user_id:
        type: integer
    required:
    - groups_id
    - user_id
    type: object
  group.UpdateGroup:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          type: integer
        type: array
      sub_groups:
        items:
          type: integer
        type: array
    required:
    - name
    type: object
...
info:
  contact:
    name: cody chen
    url: https://gbms.codychen.me/
  description: Gin swagger
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: RancoDev GBMS API Docs
  version: "1.0"
paths:
...
  /api/v1/group:
    patch:
      description: 更新群組的相關資訊
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/group.UpdateGroup'
      produces:
      - application/json
      responses:
        "204":
          description: no content
        "400":
          description: paramenters error
        "500":
          description: internal server error
      summary: 群組更新
      tags:
      - Group
    post:
      description: 建立一個群組
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/group.Group'
      produces:
      - application/json
      responses:
        "201":
          description: '{}'
        "409":
          description: create duplicates
        "500":
          description: internal server error
      summary: 創建群組
      tags:
      - Group
  /api/v1/group/{group_id}:
    get:
      description: 取得單一群組的詳細資料
      parameters:
      - description: group id
        example: 123
        in: path
        name: group_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/group.Group'
        "400":
          description: paramenters error
        "500":
          description: internal server error
      summary: 群組資訊
      tags:
      - Group
  /api/v1/group/{id}:
    put:
      description: 單一使用者改變擁有群組
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/group.GroupWindow'
      produces:
      - application/json
      responses:
        "204":
          description: no content
        "400":
          description: user not found
        "500":
          description: internal server error
      summary: 加入群組
      tags:
      - Group
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"